requirements.txt
Django==1.11.8
djangorestframework==3.7.7
model-mommy==1.5.0
pytz==2017.3
six==1.11.0
